traffy-server/server: Do not prompt error page if shaped on registration
If a user is already shaped, because he has no data volume left and is about to register, then the user will face an error message.
This is not crucial to the process itself, the registration succeeds even when prompting the error.
In order to fix this, move the shaping procedure into the exception clause.