traffy-server/web: Cast customer_id integer values to string on transfer
XMLRPC only supports 32bit unsigned integers to transfer via its RPC API. Customer_ID has an long int data type in the database and therefore exceeds this size. Workaround this by casting the integer to an string on the senders side for its transfer via RPC and casting it back to an integer on the receiver side.
TODO: Do this for all other long int/bigint fields in the databse (mostly id fields)